add notes on how to do findOrCreateUser with authCallback
